Large restructure of solution. Extracting code into files, etc.
