+++ ALsSoundSwitcher_Frontend/ALsSoundSwitcher/ALsSoundSwitcher.csproj
@@ -52,26 +52,35 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Source\ConfigIO.cs" />
-    <Compile Include="Source\IconUtils.cs" />
-    <Compile Include="Source\Settings.cs" />
-    <Compile Include="Source\Theming\LightRenderer.cs" />
-    <Compile Include="Source\Theming\DarkRenderer.cs" />
-    <Compile Include="Source\Definitions.cs" />
+    <Compile Include="Source\Logic\ClickHandlers.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Source\Theming\ThemeSwitchingLogic.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="Source\Form1.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="Source\Form1.Designer.cs">
       <DependentUpon>Form1.cs</DependentUpon>
     </Compile>
-    <Compile Include="Source\LevenshteinDistance.cs" />
+    <Compile Include="Source\Utils\ConfigUtils.cs" />
+    <Compile Include="Source\Logic\ContextMenuSetup.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Source\Utils\IconUtils.cs" />
+    <Compile Include="Source\Logic\DeviceSwitchingLogic.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Source\Utils\ProcessUtils.cs" />
+    <Compile Include="Source\Definitions\Settings.cs" />
+    <Compile Include="Source\Theming\LightRenderer.cs" />
+    <Compile Include="Source\Theming\DarkRenderer.cs" />
+    <Compile Include="Source\Definitions\Globals.cs" />
+    <Compile Include="Source\Utils\LevenshteinDistance.cs" />
     <Compile Include="Source\Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Source\Theming\CustomRenderer.cs" />
-    <EmbeddedResource Include="Source\Form1.resx">
-      <DependentUpon>Form1.cs</DependentUpon>
-      <SubType>Designer</SubType>
-    </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
@@ -82,6 +91,9 @@
       <DependentUpon>Resources.resx</DependentUpon>
       <DesignTime>True</DesignTime>
     </Compile>
+    <EmbeddedResource Include="Source\Form1.resx">
+      <DependentUpon>Form1.cs</DependentUpon>
+    </EmbeddedResource>
     <None Include="packages.config" />
     <None Include="Properties\Settings.settings">
       <Generator>SettingsSingleFileGenerator</Generator>
