+++ ALsSoundSwitcher_Frontend/ALsSoundSwitcher/Source/Definitions/Settings.cs
