+++ ALsSoundSwitcher_Frontend/ALsSoundSwitcher/Source/Utils/ConfigUtils.cs
@@ -12,7 +12,7 @@ namespace ALsSoundSwitcher
     {
       try
       {
-        var items = File.ReadAllText(Definitions.ConfigFile).Split(new[] {"\r\n", "\r", "\n"}, StringSplitOptions.None);
+        var items = File.ReadAllText(Globals.ConfigFile).Split(new[] {"\r\n", "\r", "\n"}, StringSplitOptions.None);
         foreach (var item in items)
         {
           if (item.Length == 0)
@@ -20,24 +20,24 @@ namespace ALsSoundSwitcher
             continue;
           }
 
-          var indexOfDelim = item.IndexOf(Definitions.ConfigDelimiter, StringComparison.Ordinal);
+          var indexOfDelim = item.IndexOf(Globals.ConfigDelimiter, StringComparison.Ordinal);
           var key = item.Substring(0, indexOfDelim).Trim().ToLower();
           var val = item.Substring(indexOfDelim + 1).Trim();
           dictionary.Add(key, val);
         }
 
         string buff;
-        if (TryGetValue(Definitions.ConfigKeys.BalloonTime, out buff))
+        if (TryGetValue(Globals.ConfigKeys.BalloonTime, out buff))
         {
           Settings.Current.BalloonTime = Convert.ToInt32(buff);
         }
 
-        if (TryGetValue(Definitions.ConfigKeys.BestNameMatchPercentageMinimum, out buff))
+        if (TryGetValue(Globals.ConfigKeys.BestNameMatchPercentageMinimum, out buff))
         {
           Settings.Current.BestNameMatchPercentageMinimum = Convert.ToInt32(buff);
         }
 
-        if (TryGetValue(Definitions.ConfigKeys.DarkMode, out buff))
+        if (TryGetValue(Globals.ConfigKeys.DarkMode, out buff))
         {
           Settings.Current.DarkMode = Convert.ToInt32(buff);
         }
@@ -63,11 +63,11 @@ namespace ALsSoundSwitcher
 
     private static void WriteAllSettingsToConfigFile()
     {
-      using (var sw = File.CreateText(Definitions.ConfigFile))
+      using (var sw = File.CreateText(Globals.ConfigFile))
       {
         foreach (var field in typeof(Settings.SettingsStruct).GetFields())
         {
-          var buff = field.Name + Definitions.ConfigDelimiter + field.GetValue(Settings.Current);
+          var buff = field.Name + Globals.ConfigDelimiter + field.GetValue(Settings.Current);
           sw.WriteLine(buff);
         }
       }
