+++ ALsSoundSwitcher_Frontend/ALsSoundSwitcher/Source/Utils/IconUtils.cs
