+++ ALsSoundSwitcher_Frontend/ALsSoundSwitcher/Source/Utils/LevenshteinDistance.cs
